split EnrollmentDisplay into HomeEnrollmentsDashboard + ProgramEnrollmentDisplay
EnrollmentDisplay.tsx housed both the home "My Learning" section and the
single-program dashboard behind a thin programId-based dispatch. Split
into two sibling files, drop the dispatch wrapper, and update callers
to import the dashboard they actually want.

HomeEnrollmentsDashboard.tsx owns:
  - the Wrapper / Title / EnrollmentsList styled components
  - sortEnrollments, dedupeEnrollments, getResourceKey, and the
    ProgramAsCourse helpers used only by the home dashboard
  - EnrollmentExpandCollapse and its MIN_VISIBLE constant
  - getTopLevelProgramEnrollments, getNonContractProgramEnrollments
  - the renamed component (was AllEnrollmentsDisplay)

ProgramEnrollmentDisplay.tsx owns:
  - ProgramLanguageSelect, ProgramCertificateButton (exported, still
    consumed by ProgramAsCourseCard)
  - the requirement-section types and extractResourcesFromNode /
    getRequirementSectionTitle helpers
  - the ProgramEnrollmentDisplay component

DashboardCardStyled is duplicated in both files — small styled wrapper
slated for replacement by CoursewareCard in a later phase.

EnrollmentDisplay.test.tsx splits along the same line: the top-level
home tests move to HomeEnrollmentsDashboard.test.tsx; the inner
"ProgramId Filtering" describe flattens into
ProgramEnrollmentDisplay.test.tsx. Unused imports drop from each side
(act/setupEnrollments stay home-only; waitFor/mockAxiosInstance/
invariant stay program-only). HomeContent, ProgramContent,
ProgramAsCourseCard, ProgramContent.test, and DashboardDialogs.test
update to import the new components. Zero behavior change.
